Switch GLBench glDrawElements calls to GL_UNSIGNED_SHORT.
We were using GL_UNSIGNED_INT which is not part of the standard GL ES spec.

BUG=chromium-os:32455
TEST=Same results on i915.  Sane results for affected tests on Arm.

Reviewed-on: https://gerrit.chromium.org/gerrit/28417
Tested-by: Stuart Abercrombie <sabercrombie@chromium.org>
Commit-Ready: Stuart Abercrombie <sabercrombie@chromium.org>

+++ autotest/client/deps/glbench/attributefetchtest.cc
@@ -99,7 +99,7 @@ bool AttributeFetchShaderTest::Run() {
   GLuint vertex_buffer = SetupVBO(GL_ARRAY_BUFFER,
                                   vertex_buffer_size, vertices);
 
-  GLuint *indices = NULL;
+  GLushort *indices = NULL;
   GLuint index_buffer = 0;
   GLsizeiptr index_buffer_size = 0;
